+++ src/buffer/message_view.rs
@@ -1,6 +1,6 @@
 use chrono::{TimeDelta, Utc};
-use data::config::buffer::Dimmed;
 use data::config::buffer::nickname::ShownStatus;
+use data::config::buffer::{CondensationIcon, Dimmed};
 use data::isupport::{CaseMap, PrefixMap};
 use data::server::Server;
 use data::user::ChannelUsers;
@@ -67,6 +67,14 @@ pub struct ChannelQueryLayout<'a> {
 }
 
 impl<'a> ChannelQueryLayout<'a> {
+    fn condensation_marker(&self, marker: Marker) -> Marker {
+        match self.config.buffer.server_messages.condense.icon {
+            CondensationIcon::None => Marker::None,
+            CondensationIcon::Chevron => marker,
+            CondensationIcon::Dot => Marker::Dot,
+        }
+    }
+
     fn format_timestamp(
         &self,
         message: &'a data::Message,
@@ -413,7 +421,7 @@ impl<'a> ChannelQueryLayout<'a> {
         let marker = message_marker(
             if message.expanded {
                 if message.condensed.is_some() {
-                    Marker::Contract
+                    self.condensation_marker(Marker::Contract)
                 } else {
                     Marker::None
                 }
@@ -510,7 +518,7 @@ impl<'a> ChannelQueryLayout<'a> {
         let moved_link = link.clone();
 
         let marker = message_marker(
-            Marker::Expand,
+            self.condensation_marker(Marker::Expand),
             right_aligned_width,
             self.config,
             theme::selectable_text::condensed_marker,